Reorganize folder structure and move files.
