+++ src/app/app.routes.ts
@@ -4,9 +4,8 @@ import { Routes } from '@angular/router';
 import { AuthLayoutComponent } from './shared/auth-layout/auth-layout.component';
 import { SignInComponent } from './pages/sign-in/sign-in.component';
 import { BaseLayoutComponent } from './shared/base-layout/base-layout.component';
-import { authGuard } from './shared/guards/auth.guard';
+import { authGuard } from './guards/auth.guard';
 import { RolesListComponent } from './pages/roles-list/roles-list.component';
-import { ToponymsListComponent } from './pages/toponyms/toponyms-list/toponyms-list.component';
 import { UsersListComponent } from './pages/users-list/users-list.component';
 import { CountriesListComponent } from './pages/toponyms/countries-list/countries-list.component';
 import { LocalitiesListComponent } from './pages/toponyms/localities-list/localities-list.component';
