+++ src/app/pages/toponym-details/toponym-details.component.css
