+++ src/app/pages/toponyms-lists/districts-list/districts-list.component.ts
