+++ src/app/pages/toponyms-lists/localities-list/localities-list.component.ts
