+++ src/app/pages/toponyms-lists/regions-list/regions-list.component.ts
