+++ src/app/pages/toponyms-lists/toponyms-list/toponyms-list.component.css
