+++ src/app/pages/user-details/cause-of-blocking-dialog/cause-of-blocking-dialog.component.css
