--- src/app/shared/dialogs/create-role-dialog/create-role-dialog.component.css
@@ -1,8 +1,0 @@
-.full-width {
-  width: 100%;
-}
-
-/* .mdc-text-field--filled:not(.mdc-text-field--disabled)  {
-  background-color: #f9faf4 !important;
-}
- */
