--- src/app/shared/dialogs/create-role-dialog/create-role-dialog.component.html
@@ -1,24 +1,0 @@
-<h2 mat-dialog-title>Новая роль</h2>
-<p-confirmdialog />
-<p-toast position="center" />
-<mat-dialog-content class="full-width">
-  <form>
-    <mat-form-field class="full-width">
-      <mat-label>Название</mat-label>
-      <input matInput [formControl]="roleName">
-      @if (roleName.hasError('required')) {
-      <mat-error>Поле 'Название' не может быть пустым!</mat-error>
-      }
-    </mat-form-field>
-    <mat-form-field class="full-width">
-      <mat-label>Описание</mat-label>
-      <textarea rows="5" matInput [formControl]="roleDescription"></textarea>
-    </mat-form-field>
-  </form>
-</mat-dialog-content>
-
-<mat-dialog-actions>
-  <button class="dialog-cancel-button" mat-button (click)="onCancelClick($event)">Отменить</button>
-  <button class="dialog-submit-button" mat-button [disabled]="roleName.hasError('required')"
-    (click)="onCreateRoleClick()">Сохранить</button>
-</mat-dialog-actions>
