+++ src/app/shared/dialogs/details-dialogs/details-dialog/outdated-item-menu/outdated-item-menu.component.ts
